Fixes home directory slow saving
The configured BufWritePost autocmd refreshes the current Tree
and all of its descendant nodes.

This is fine in actual git projects but makes file writes really slow
when in my home directory (it recurses through all of the nodes (quite a
few of them if you have node_modules installed).

In this commit I'm updating the behaviour to only refresh git in git projects

To speed up saving files in not git projects

- [x] I've tested that it's fast on non-git project trees
- [x] I've tested that this still updates the icons for git project trees