Add a mode that allows edit without the picker
I would like to open new files in the most left window by default, but
have the option to select a specific window when needed. Currenlty the
window picking logic is either enabled or disabled, but adding this mode
allows different keybindings for both situations.

This allows me to choose to either open it in the default windows with a
single keypress (which is often my default behavior), or to open it with
the picker asking me where I want it to be opened.

Curious what you think of this. Name of the mode can of course be
changed to whatever you feel is a good name, but I would really like to
be able to use this functionality. Thanks!