fix(view): do not close window when it is replaced
With `actions.open_file` enabled, when an NvimTree buffer is changed to
some other buffer, the window in which that buffer swap occurs is
closed.

Since the `quit_on_open` function runs on `NvimTree` `BufWipeout`, there
should not be a need to close the window that holds that buffer.

This issue could be reproduced by:

1. Opening an existing file `:e firstfile.txt`
2. `:sp`
3. `:lua require("nvim-tree").open_replacing_current_buffer()`
4. Replacing the NvimTree buffer with some other buffer by
   `:e somefile.txt`

The window containing the buffer opened in the first step would be
closed. There would be only a single window with `somefile.txt` on that
tab page.

With this commit, that window is retained and the tab now contains
`firstfile.txt` in the left window and `somefile.txt` in the right
window.