feat(bookmarks): add bookmark feature for nodes
Uses the signcolum to display the mark status.
Marks are saved in memory for now, we'll see if we want to implement a
filesystem save for the marks (would not be hard).
Added `m` keybinding to toggle the mark on the node.
Users can call `require "nvim-tree.marks".get_marks()` which returns a
list of absolute paths.