abort iteration when fs-event-incapable
Fixes #2866. The original code, while detecting whether Windows NT
allows the dir to be iterated through, will go to the other branch
of the if...else... statement, which is intended for cases when
filter_reason is not none. Simply moving `is_fs_event_capable` into the
first branch should solve the problem.