fix: api.tree.change_root_to_node on a file now changes directory to parent as per documentation (#3228)
fix: wrong behavior for api.tree.change_root_to_node

As described in the documentation, `api.tree.change_root_to_node` should
set the tree root to the parent directory when the target node is a file.

Currently, calling this function on a file node does nothing.
This commit adds logic to handle cases where the input node is a `FileNode`
in the `change_root_to_node` function.

Co-authored-by: Alexander Courtis <alex@courtis.org>